api/requirements.txt
black==23.7.0
blinker==1.6.2
click==8.1.7
colorama==0.4.6
Flask==2.3.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.0.5
Flask-SocketIO==5.3.6
geopy==2.4.1
greenlet==2.0.2
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
mailjet-rest==1.3.4
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-sqlalchemy==1.0.0
mypy-extensions==1.0.0
OpenAI==1.14.3
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
pytest-cov==5.0.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
PyPDF2==3.0.1
pytest==8.1.1
python-dotenv==1.0.0
python-socketio==5.11.2  # Required by Flask-SocketIO
python-engineio==4.9.0  # Required by Flask-SocketIO
SQLAlchemy==2.0.20
typing_extensions==4.7.1
Werkzeug==2.3.7
pusher==3.3.2
